refactor(core/presentation): Move some code around - Extract evaluateExpression to a separate file - Move FormField and FormikFormField to parent directory
