+++ app/scripts/modules/core/src/presentation/forms/fields/index.ts
@@ -1,4 +1,2 @@
 export * from './FormikExpressionField';
 export * from './FormikExpressionRegexField';
-export * from './FormField';
-export * from './FormikFormField';
