+++ app/scripts/modules/core/src/presentation/forms/index.ts
@@ -1,4 +1,6 @@
+export * from './FormField';
 export * from './FormikForm';
+export * from './FormikFormField';
 export * from './Validation';
 export * from './fields';
 export * from './inputs';
