+++ app/scripts/modules/core/src/presentation/forms/inputs/expression/index.ts
@@ -1,6 +1,7 @@
 ///<reference path=".//spel2js.d.ts" />
 
-export * from './ExpressionInput';
 export * from './ExpressionError';
+export * from './ExpressionInput';
 export * from './ExpressionPreview';
+export * from './evaluateExpression';
 export * from './spel2js.templateParser';
